+++ signalr-go-server/signalr/clientproxy.go
