+++ signalr-go-server/signalr/groupmanager.go
