+++ signalr-go-server/signalr/handshake.go
