+++ signalr-go-server/signalr/hublifetimemanager.go
