+++ signalr-go-server/signalr/jsonhubprotocol.go
