+++ signalr-go-server/signalr/streamclient.go
